Changes to accomodate previous tweaks.
The underline under the h2 can overlap with the aside. Changing the display of the header to grid seems to fix that.

Removing the top margin from paragraphs meant that they were shoved up against the bottom of the details elements.
Details were given a bottom margin to compensate.
Incidentally, this fixes a minor issue with the McCall job search page. But introduces issues into the wordle page.