New main width paradigm to enable full bleed
I want to have elements inside an article that spans the entire width of the page.
The vw unit seemed ideal for this. But alas, browsers incorrectly include the width of the scroll bar
So that won't work.
("sleek" front end design ruining functionality, as is typical)

The workaround is to make the main really wide, and then restrict the width of all but a few children inside of it.
From:
https://kilianvalkhof.com/2020/css-html/full-bleed-layout-using-simple-css/

The +rem on the width is to cancel out the padding in the container above main.

I am very concerned about whether this will break the centering of elements like katex equations.
Oh well! There's an easy way to find out!