Update head.html - use beta mathjax v4
The main thing of note in v4 is that linebreaks are actually supported again. I looked at the workarounds in v3 and none of them would preserve both appearance and compatibility with katex. So fingers crossed that beta v4 of mathjax works. If it has any major issues, I'll need to revert back to katex.