Update compiled-jtd-style.css
The kgjs (kineticgraphs javascript) library loads a LOT of css. It first loads a copy of the katex css, then it loads the contents of [repo]/build/kg.css, whose contents in turn are a mish-mash of (things vital for certain features of kgjs) and (random bits of styling for the econgraphs website). Loading a kgjs graph makes the scrollbar dissapear, for example.

I'm going to remove the kgjs css from my notes, and I think this is the most vital part of the katex css to prevent duplicate text draws. (I could also just always load the katex css, but this seems cleaner.)