Update compiled-jtd-style.css
Alas, I was incorrect that the mathml styling would be enough to prevent rendering errors. As a stop-gap, I am importing the entirety of the katex css on any page where I has a kg-container. I've left this snippet here as a reminder to come back and try again later to extract the bare minumum css to make katex function. (probably not even worth the time, if I'm being honest; it's like 30kb of css. That's... fine...ish